+++ src/entities/product-required-option.entity.ts
@@ -2,7 +2,7 @@ import { Entity, Column, ManyToOne, JoinColumn, OneToMany } from 'typeorm';
 import { CommonEntity } from './common/common.entity';
 import { ProductEntity } from './product.entity';
 import { ProductInputOptionEntity } from './product-input-option.entity';
-import { CartRequiredOptionEntity } from './cart-product-required-option.entity';
+import { CartRequiredOptionEntity } from './cart-required-option.entity';
 
 @Entity({ name: 'product_required_option' })
 export class ProductRequiredOptionEntity extends CommonEntity {
